// src/app/backend/model/adminHuntRequestDto.ts
/**
 * OutOut API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminHuntGoalRequestDto } from './adminHuntGoalRequestDto';


export interface AdminHuntRequestDto { 
    id?: number | null;
    name?: string | null;
    city?: string | null;
    cityPart?: string | null;
    startAt?: string;
    endAt?: string;
    goals?: Array<AdminHuntGoalRequestDto> | null;
}